Land Hydrology

Domains

LandHydrology.Domains.Column — Type

Column{FT} <: AbstractVerticalDomain

A struct holding the necessary information to construct a domain, a mesh, a center and face space, etc. For use when a finite difference in 1D is suitable.

Fields

zlim

Domain interval limits, (zmin, zmax)

nelements

Number of elements used to discretize the interval

boundary_tags

Boundary face identifiers

LandHydrology.Domains.make_function_space — Function

make_function_space(domain::Column)

Returns the center and face space z values of the column domain.

LandHydrology.Domains.AbstractVerticalDomain — Type

AbstractVerticalDomain{FT}

An abstract type for vertical domains, using the floating point type FT.

Models

LandHydrology.Models.AbstractModel — Type

abstract type AbstractModel

An abstract type for models.

Eventually, the land model and all major subcomponents will be of this type.