Simulations Interface

Models

LandHydrology.Simulations.Simulation — Type

struct Simulation <: AbstractSimulation

A simulation wraps an abstract LandHydrology model containing equation specifications and an instance of an integrator used for time integration of the discretized model PDE.

Fields

model

A LandHydrology model

integrator

a DiffEqBase.jl integrator used for time

callbacks

user defined callback operations

Simulation Execultion

LandHydrology.Simulations.run! — Function

run!(simulation::AbstractSimulation, args...; kwargs...)

Run a simulation to the end.

SciMLBase.step! — Function

step!(simulation::AbstractSimulation, args...; kwargs...)

Step forward a simulation one time step.