Soil Interface

Parameters

LandHydrology.SoilInterface.SoilParams — Type

SoilParams{FT} <: ParameterStructure{FT}

Defaults correspond to loam soil.

Fields

ν

Porosity

S_s

Specific storage

ν_ss_gravel

Volumetric fraction of soil solids in gravel

ν_ss_om

Volumetric fraction of soil solids in organic matter

ν_ss_quartz

Volumetric fraction of soil solids in quartz/sand

ρc_ds

Volumetrcy heat capacicity of dry soil

κ_solid

Thermal conductivity of soil solids

ρp

Particle density

κ_sat_unfrozen

Thermal conductivity of saturated unfrozen soil

κ_sat_frozen

Thermal conductivity of saturated frozen soil

a

Parameter needed in Balland & Arp model for Kersten number

b

Parameter needed in Balland & Arp model for Kersten number

κ_dry_parameter

Parameter needed in Balland & Arp model for κ_dry

z_0m

Surface roughness for momentum

z_0s

Surface roughness for scalars

Models

LandHydrology.SoilInterface.SoilModel — Type

SoilModel{FT, dm, em <: AbstractSoilModel, hm <: AbstractSoilModel, bc, sp,ep,n}

The model type for the soil model.

Fields

domain

energy_model

Soil energy model - prescribed or dynamics

hydrology_model

Soil hydrology model - prescribed or dynamic

boundary_conditions

Boundary conditions tuple

soil_param_set

Soil parameters

earth_param_set

Earth parameter set

name

name

LandHydrology.SoilInterface.SoilHydrologyModel — Type

SoilHydrologyModel <: AbstractSoilComponentModel

The model type to be used when the user wants to simulate the flow of water in soil by solving Richards equation.

Fields

hydraulic_model

viscosity_factor

impedance_factor

LandHydrology.SoilInterface.SoilEnergyModel — Type

SoilEnergyModel <: AbstractSoilComponentModel

The model type to be used when the user wants to simulate heat transfer in soil by solving the heat partial differential equation.

Fields

LandHydrology.SoilInterface.PrescribedTemperatureModel — Type

PrescribedTemperatureModel <: AbstractSoilComponentModel

The model type to be used when the user does not wish to solve the heat partial differential equation, but instead wishes to prescibe a temperature profile in the soil.

This is useful for situations where Richards Equation alone is sufficient. Because the hydraulic conductivity can be a function of temperature, a temperature profile can be supplied in order to simulate that. The default is 288K across the domain, the reference temperature for the viscosity effect.

Fields

T_profile

Profile of (z,t) for temperature

LandHydrology.SoilInterface.PrescribedHydrologyModel — Type

PrescribedHydrologyModel <: AbstractSoilComponentModel

The model type to be used when the user does not wish to solve the Richards equation, but instead wishes to prescibe a water content profile in the soil.

This is useful for situations where only the heat equation is to be solved. Because the thermal conductivity and heat capacities depend on water content, a water profile must be defined to solve the heat equation. The default for both ice and liquid water content is zero, which applies for totally dry soil.

Fields

ϑ_l_profile

Profile of (z,t) for ϑ_l

θ_i_profile

Profile of (z,t) for θ_i

Right hand side

LandHydrology.SoilInterface.make_rhs — Function

make_rhs(model::SoilModel)

A function which takes a model::AbstractModel as argument, and returns function which computes the rhs of a set of ordinary differential equations corresponding to that model.

Currently, the arguments of the returned rhs function are configured for use with OrdinaryDiffEq.jl. For the soil model, the rhs function depends on the type of the components of the model (the energy and hydrology models), as well as whether additional sources are included.

make_rhs(energy::PrescribedTemperatureModel, hydrology::PrescribedHydrologyModel, model::SoilModel)

make_rhs(energy::PrescribedTemperatureModel, hydrology::SoilHydrologyModel, model::SoilModel)

make_rhs(energy::SoilEnergyModel, hydrology::PrescribedHydrologyModel, model::SoilModel)

make_rhs(energy::SoilEnergyModel, hydrology::SoilHydrologyModel, model::SoilModel)

Boundary Conditions

LandHydrology.SoilInterface.VerticalFlux — Type

VerticalFlux{f <: AbstractFloat} <: AbstractBC

The BC type to be used for prescribed vertical boundary fluxes. The flux is assumed to be of the form

$F = f ẑ$

where f is the value supplied by the user (currently a constant).

Fields

flux

Scalar flux; positive = aligned with ẑ

LandHydrology.SoilInterface.NoBC — Type

NoBC <: AbstractBC

The BC type to be used when equations do not require boundary conditions, e.g. for Prescribed Models.

LandHydrology.SoilInterface.FreeDrainage — Type

FreeDrainage <: AbstractBC

A BC type for use with the SoilHydrologyModel (Richards Equation), at the bottom of the domain, setting

$∇h = 1.$

LandHydrology.SoilInterface.Dirichlet — Type

Dirichlet{f} <: AbstractBC

A BC type setting a boundary value of the state ϑ_l or temperature T. This can be a function of time.

Note that we apply boundary conditions to temperature even though volumetric internal energy is the prognostic variable.

Fields

state_value

State Value; (t) -> f(t)

Missing docstring.

Missing docstring for SoilDomainBC. Check Documenter's build log for details.

LandHydrology.SoilInterface.SoilComponentBC — Type

SoilComponentBC{ebc <: AbstractBC, hbc <: AbstractBC} <: AbstractFaceBC

A container for holding the boundary conditions for the components of the soil model.

The values must be of type AbstractBC; the two components are energy and hydrology. Each boundary will have a SoilComponentBC object associated with it.

Fields

energy

BC for the heat equation

hydrology

BC for ϑ_l

LandHydrology.SoilInterface.PrescribedAtmosForcing — Type

PrescribedAtmosForcing{FT <: AbstractFloat} <: AbstractFaceBC

A container for holding the atmosphere state variables that are needed to drive the soil model via a Monin-Obukhov theory for the surface fluxes.

At present, this boundary condition can only be used with a SoilModel that includes both volumetric internal energy and volumetric water fractions as prognostic variables. Sublimation of ice is not currently supported, nor is precipitation, nor any additional resistance due to a dry layer in the soil. Eventually this will support functions of space and time, so that reanalysis data can be used to drive the land model.

Fields

u_atm

The wind speed parallel to the surface at height z_atm

θ_atm

The potential temperature of the atmosphere at the height z_atm

z_atm

The height at which the atmospheric state is measured

θ_scale

The porential temperature scale

ρ_a_sfc

The density of the moist air at the surface

q_atm

The specific humidity of the atmosphere at the height z_atm

LandHydrology.SoilInterface.boundary_fluxes — Function

boundary_fluxes(
    X,
    bc::SoilComponentBC,
    face::Symbol,
    model::SoilModel,
    cs,
    t,
)

Returns the boundary flux at the boundary face for all components of the soil model.

X is a FieldVector with components ϑ_l, θ_i, and T. The bc field contains the boundary condition types (Dirichlet, VerticalFlux, etc) for both the energy and hydrology equations. In the case of explicitly time-varying boundary conditions, we also must supply the current time t.

The remaining fields are required in the case of certain bc types, such as Dirichlet. In this case, we are turning a state boundary condition into a flux boundary condition, and this requires knowing the side of the domain we are on (encoded in face), the values of X just interior to this face, and the center space cs itself.

function boundary_fluxes(
    X,
    bc::PrescribedAtmosForcing,
    face::Symbol,
    model::SoilModel,
    cs,
    t,
)

Returns the boundary flux at the boundary face for all components of the soil model in the case of PrescribedAtmosForcing.

X is a FieldVector with components ϑ_l, θ_i, and T. The bc field contains the boundary condition types (Dirichlet, VerticalFlux, etc) for both the energy and hydrology equations. In the case of explicitly time-varying boundary conditions, we also must supply the current time t.

The remaining fields are required in the case of certain bc types, such as Dirichlet. In this case, we are turning a state boundary condition into a flux boundary condition, and this requires knowing the side of the domain we are on (encoded in face), the values of X just interior to this face, and the center space cs itself.

Missing docstring.

Missing docstring for compute_turbulent_boundary_fluxes. Check Documenter's build log for details.

Initial Conditions

Missing docstring.

Missing docstring for set_initial_state. Check Documenter's build log for details.

Missing docstring.

Missing docstring for default_initial_conditions. Check Documenter's build log for details.