Soil Water Parameterizations

LandHydrology.SoilInterface.SoilWaterParameterizations.vanGenuchten — Type

vanGenuchten{FT} <: AbstractHydraulicsModel{FT}

The necessary parameters for the van Genuchten hydraulic model; defaults are for loam, setting the residual water fraction to zero.

Fields

n

Pore size distribution index, unitless

α

Inverse of the air entry potential (1/m)

m

Exponent parameter - determined as 1-1/n

θr

Residual water content

Ksat

Saturated hydraulic conductivity, m/s

LandHydrology.SoilInterface.SoilWaterParameterizations.volumetric_liquid_fraction — Function

volumetric_liquid_fraction(
    ϑ_l::FT,
    ν_eff::FT,
) where {FT}

Computes the volumetric liquid fraction from the effective porosity and the augmented liquid fraction.

LandHydrology.SoilInterface.SoilWaterParameterizations.effective_saturation — Function

effective_saturation(porosity::FT, ϑ_l::FT, θr::FT)

Computes the effective saturation of the soil, given the augmented liquid fraction ϑ_l, the residual water fraction θr, and a porosity.

This porosity can be effective (ν-θ_i) or true (ν) depending on the application. Note that ϑl is defined to be larger than θr. If `ϑl-θr` is negative, hydraulic functions that take it as an argument will return imaginary numbers, resulting in domain errors.

LandHydrology.SoilInterface.SoilWaterParameterizations.pressure_head — Function

pressure_head(hm::vanGenuchten{FT}, ϑ_l::FT, ν_eff::FT, S_s::FT)

Computes the pressure head according to the hydraulics model hm.

For unsaturated soil (ϑ_l < ν_eff, or S_l_eff < 1), the matric potential is used. It is evaluated at an effective saturation computed using the effective porosity. This ensures that the pressure head is continuous across the saturated and unsaturated border. For saturated soil, the pressure head is positive.

LandHydrology.SoilInterface.SoilWaterParameterizations.hydraulic_conductivity — Function

hydraulic_conductivity(hm::vanGenuchten{FT}, S::FT, viscosity_f::FT, impedance_f::FT) where {FT}

Computes the hydraulic conductivity given a hydraulics model hm, the effective saturation, and the additional viscosity and ice impedance factors, which must be computed prior to evaluating this function.

LandHydrology.SoilInterface.SoilWaterParameterizations.hydrostatic_profile — Function

hydrostatic_profile(hm, z, zmin, ν)

Returns the augmented liquid fraction corresponding to a hydrostatic profile in variably saturated soil, with the boundary between the saturated and unsaturated zones at z_∇.

LandHydrology.SoilInterface.SoilWaterParameterizations.matric_potential — Function

matric_potential(hm::vanGenuchten{FT}, S)

Computes the matric potential according the the hydraulics model hm, given an effective saturation S.

LandHydrology.SoilInterface.SoilWaterParameterizations.inverse_matric_potential — Function

inverse_matric_potential(
    hm::vanGenuchten{FT},
    ψ::FT
) where {FT}

Compute the effective saturation given the matric potential ψ, using the van Genuchten formulation hm.

LandHydrology.SoilInterface.SoilWaterParameterizations.viscosity_factor — Function

viscosity_factor(
    vm::NoEffect{FT},
    _...,
) where {FT}

Returns a multiplicative factor of 1.0, which does not account for the temperature dependence of the conductivity.

viscosity_factor(
    vm::TemperatureDependentViscosity{FT},
    T::FT,
) where {FT}

Returns the multiplicative factor which accounts for the temperature dependence of the conductivity.

LandHydrology.SoilInterface.SoilWaterParameterizations.impedance_factor — Function

impedance_factor(
    imp::NoEffect{FT},
    _...,
) where {FT}

Returns a multiplicative factor of 1.0, which does not account for the ice on conductivity.

impedance_factor(
    imp::IceImpedance{FT},
    f_i::FT,
) where {FT}

Returns the multiplicative factor when an effect due to the fraction of ice f_i is desired.

LandHydrology.SoilInterface.SoilWaterParameterizations.AbstractConductivityFactor — Type

AbstractConductivityFactor{FT <: AbstractFloat}

An abstract type for multiplicative factors that affect the hydraulic conductivity, increasing or decreasing it relative to the formula that depends only on liquid water content.

Examples include a multiplicative factor due to ice impedance or one due to the temperature dependence of the viscosity of water.

LandHydrology.SoilInterface.SoilWaterParameterizations.IceImpedance — Type

IceImpedance{FT} <: AbstractConductivityFactor{FT}

The factor type which alters the conductivity to account for the effects of ice in the pore space.

The empirical factor comes from Lundin (1990).

LandHydrology.SoilInterface.SoilWaterParameterizations.TemperatureDependentViscosity — Type

TemperatureDependentViscosity{FT} <: AbstractConductivityFactor{FT}

The factor type which alters the conductivity to account for the effects of temperature on the viscosity of water.

LandHydrology.SoilInterface.SoilWaterParameterizations.NoEffect — Type

NoEffect{FT} <: AbstractConductivityFactor{FT}

The default type of conductivity factor, used to set a particular multiplicative factor to unity.