Soil Interface

Parameters

LandHydrology.SoilInterface.SoilParams — Type

struct SoilParams{FT} <: ParameterStructure{FT}

Models

LandHydrology.SoilInterface.SoilModel — Type

SoilModel{FT, domain, em <: AbstractSoilModel, hm <: AbstractSoilModel, bc, A,B}

The model type for the soil model.

Fields

domain

energy_model

Soil energy model - prescribed or dynamics

hydrology_model

Soil hydrology model - prescribed or dynamic

boundary_conditions

Boundary conditions tuple

soil_param_set

Soil parameters

earth_param_set

Earth parameter set

name

name

variables

variables

LandHydrology.SoilInterface.SoilHydrologyModel — Type

SoilHydrologyModel

The model type to be used when the user wants to simulate the flow of water in soil by solving Richards equation.

Fields

hydraulic_model

LandHydrology.SoilInterface.SoilEnergyModel — Type

SoilEnergyModel

The model type to be used when the user wants to simulate heat transfer in soil by solving the heat partial differential equation.

Fields

LandHydrology.SoilInterface.PrescribedTemperatureModel — Type

Base.@kwdef struct PrescribedTemperatureModel <: AbstractSoilComponentModel
    "Profile of (z,t) for temperature"
     T_profile::Function = (z,t) -> eltype(z)(288)
end

The model type to be used when the user does not wish to solve the heat partial differential equation, but instead wishes to prescibe a temperature profile in the soil.

This is useful for situations where Richards Equation alone is sufficient. Because the hydraulic conductivity can be a function of temperature, a temperature profile can be supplied in order to simulate that. The default is 288K across the domain, the reference temperature for the viscosity effect.

Fields

T_profile

Profile of (z,t) for temperature

LandHydrology.SoilInterface.PrescribedHydrologyModel — Type

Base.@kwdef struct PrescribedHydrologyModel <: AbstractSoilComponentModel
    "Profile of (z,t) for ϑ_l"
    ϑ_l_profile::Function = (z,t) -> eltype(z)(0.0)
    "Profile of (z,t) for θ_i"
    θ_i_profile::::Function = (z,t) -> eltype(z)(0.0)
end

The model type to be used when the user does not wish to solve the Richards equation, but instead wishes to prescibe a water content profile in the soil.

This is useful for situations where only the heat equation is to be solved. Because the thermal conductivity and heat capacities depend on water content, a water profile must be defined to solve the heat equation. The default for both ice and liquid water content is zero, which applies for totally dry soil.

Fields

ϑ_l_profile

Profile of (z,t) for ϑ_l

θ_i_profile

Profile of (z,t) for θ_i

Right hand side

LandHydrology.SoilInterface.make_rhs — Function

make_rhs(model::SoilModel)

A function which takes a model::AbstractModel as argument, and returns function which computes the rhs of a set of ordinary differential equations corresponding to that model.

Currently, the arguments of the returned rhs function are configured for use with OrdinaryDiffEq.jl. For the soil model, the rhs function depends on the type of the components of the model (the energy and hydrology models), as well as whether additional sources are included.

Boundary Conditions

LandHydrology.SoilInterface.VerticalFlux — Type

VerticalFlux{f <: AbstractFloat} <: AbstractBC

The BC type to be used for prescribed vertical boundary fluxes. The flux is assumed to be of the form

$F = f ẑ$

where f is the value supplied by the user (currently a constant).

Fields

flux

Scalar flux; positive = aligned with ẑ

LandHydrology.SoilInterface.NoBC — Type

NoBC <: AbstractBC

The BC type to be used when equations do not require boundary conditions, e.g. for Prescribed Models.

LandHydrology.SoilInterface.SoilDomainBC — Type

struct SoilDomainBC{D, TBC, BBC}

A container holding the SoilComponentBC for each boundary face.

Each field value should be of type SoilComponentBC. This doesn't do what we want. Ideally the fields would change depending on the domain - e.g. for a Column, they are top and bottom, but for a 3D domain, they might be top, bottom, xleft, xright, yleft, yright.

Fields

top

SoilComponentBC for the top of the domain

bottom

SoilComponentBC for the bottom of the domain

LandHydrology.SoilInterface.SoilComponentBC — Type

struct SoilComponentBC{ebc <: AbstractBC, hbc <: AbstractBC}

A container for holding the boundary conditions for the components of the soil model.

The values must be of type AbstractBC; the two components are energy and hydrology. Each boundary will have a SoilComponentBC object associated with it.

Fields

energy

BC for the heat equation

hydrology

BC for ϑ_l

Missing docstring.

Missing docstring for compute_vertical_flux. Check Documenter's build log for details.

Initial Conditions

LandHydrology.SoilInterface.set_initial_state — Function

set_initial_state(model::SoilModel, f::Function, t0::Real)

A function which sets the initial state for the soil model, given a function of space (and time) f, as well as the initial time t0.